+++ apps/intra/src/shared/components/ui/header/index.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import { useState } from 'react';
+import { useAuthStore } from '../../../store/auth-store';
+import { MobileHeader } from './mobile-header';
+import { DesktopHeader } from './desktop-header';
+
+export default function Header(): React.ReactElement {
+  const [isMobileSearchOpen, setIsMobileSearchOpen] = useState(false);
+  const { isAuthenticated } = useAuthStore();
+
+  return (
+    <header className="flex w-full items-center justify-between border-b border-gray-200">
+      <div className="mx-auto flex w-full max-w-[1200px] items-center justify-between px-5 py-4">
+        <MobileHeader
+          isAuthenticated={isAuthenticated}
+          isMobileSearchOpen={isMobileSearchOpen}
+          setIsMobileSearchOpen={setIsMobileSearchOpen}
+        />
+        <DesktopHeader isAuthenticated={isAuthenticated} />
+      </div>
+    </header>
+  );
+}
